added functions to finish out the edit tab
added undo, redo, cut, copy, paste functions. 

still need to implement the save function that overwrites the old version of the file with the new version.
still need to implement the new file function that prompts to save or not, and then clears the window and opens a new file.
still need to add themes.