implemented new_file() and a buggy save_file()
new_file() works as expected

I spent today getting bogged down in documentation, trying to figure out how to "grab" the working file name when a file has been "saved-as" or "opened". 
I've implemented a boolean class attribute for self.file_name_set to handle the first part of the save function. 
If a file name has not been set, then the file is new. it has not been saved as or opened from storage. So if the user wants to save the file, a "save as" dialog box appears. 
This is all fine and good.
When a file name HAS been set, however (a file has been opened from storage, or previously "saved as"), I'd like to instantiate another class attribute (self.filename) with the file's name. 
This way, if a filename has been set, the save function can overwrite the previous version of the saved file, rather than prompting a new "save-as" dialog box. 

I may abandon this approach if future attempts continue coming up empty, and get rid of the "save" option entirely, and rely on the "save as" button for all saving activity. Not optimal, but functions okay, as long as the user remembers to "save as" frequently, and doesn't get annoyed by popup boxes...