added themes, and other fun stuff
added 10 custom themes and set the default theme to light!
currently, only supporting courier new font, and only in one (variable) size for the entire document, as this is meant to be a very barebones text editor... basically a software typewriter. 
May add other fonts in the future, but will keep the one font per document as a standard.

also fixed a missing boolean assignment when a new file is created. 
Now, self.file_name_set() is changed to false when a new file is opened, as should be the case.

future iterations could include word count functionality, shortcut keys (cmd + s to save, etc), and, as mentioned above, support for other fonts.