added status bar & word count placeholder
added a status bar, and a word count placeholder. it's an f-string that shows "word count: " followed by the word count as an int.
I adjusted all of the themes to show the status bar text in the same color as the body text

I wrote a placeholder function for updating the word count, but I'll need to revisit this later, as it's currently just displaying the word count of zero.

I need to come up with the word counting logic, and implement this to update the self.word_count attribute, and then show a new message in the status bar that displays the updated word count. Shouldn't be too difficult, just didn't have time for it yet :)