added ability to set default theme and more
I added an option within the themes menu to either select the theme for immediate use, or to set it as the default theme, which will load upon running the program.
I achieved this by instantiating a text file (default_theme.txt) that stores the name of the selected default theme as a string. The default is light, but whenever the user selects a new default theme, the text in default_theme.txt is replaced with the name of the theme they selected. This file is read upon program launch and whichever theme is saved here will be loaded immediately.

I also changed the font of the word count in the status bar to match the font of the text window.